cache implementation

+++ src/main/java/com/example/SpringSecurity/repository/OutboxRepository.java
@@ -3,6 +3,6 @@ package com.example.SpringSecurity.repository;
 import com.example.SpringSecurity.entity.OutboxTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OutboxRepository extends JpaRepository<OutboxTable,Integer> {
+public interface OutboxRepository extends JpaRepository<OutboxTable,Long> {
 
 }
